refactor: move plugin management to dedicated controller
