+++ BTCPayServer/Components/GlobalNav/Default.cshtml
@@ -12,7 +12,7 @@
 @model BTCPayServer.Components.GlobalNav.GlobalNavViewModel
 
 @{
-    var pluginsUrl = Url.Action("ListPlugins", "UIServer");
+    var pluginsUrl = Url.Action("ListPlugins", "UIPluginManager");
     var loginName = User.Identity?.Name;
     var activePage = ViewData["ActivePage"]?.ToString();
     var serverEmailsPage = $"Server-{nameof(ServerNavPages.Emails)}";
@@ -50,10 +50,10 @@
                 @if (!string.IsNullOrEmpty(pluginsUrl))
                 {
                     <li class="py-1 px-3">
-                        <a class="dropdown-item nav-link" asp-area="" asp-controller="UIServer" asp-action="ListPlugins" text-translate="true">Installed Plugins</a>
+                        <a class="dropdown-item nav-link" asp-area="" asp-controller="UIPluginManager" asp-action="ListPlugins" text-translate="true">Installed Plugins</a>
                     </li>
                     <li class="py-1 px-3">
-                        <a class="dropdown-item nav-link" asp-area="" asp-controller="UIServer" asp-action="PluginDirectory" text-translate="true">Plugin Directory</a>
+                        <a class="dropdown-item nav-link" asp-area="" asp-controller="UIPluginManager" asp-action="PluginDirectory" text-translate="true">Plugin Directory</a>
                     </li>
                 }
             </ul>
