+++ BTCPayServer/Hosting/BTCPayServerServices.cs
@@ -298,7 +298,6 @@ namespace BTCPayServer.Hosting
             services.AddExceptionHandler<PluginExceptionHandler>();
             services.TryAddSingleton<AppService>();
             services.AddTransient<PluginService>();
-            services.AddTransient<PluginManagementProjectionService>();
             services.AddSingleton<PluginHookService>();
             services.AddSingleton<IPluginHookService, PluginHookService>(provider => provider.GetService<PluginHookService>());
             services.TryAddTransient<Safe>();
