+++ BTCPayServer/Plugins/Translations/Translations.Default.cs
@@ -48,6 +48,7 @@ namespace BTCPayServer.Plugins.Translations
   "{0} Wallet": "",
   "{0} Wallet Labels": "",
   "{0} Wallet Settings": "",
+  "{0} details": "",
   "{0}-of-{1} {2} · {3}/{4} submitted · {5}": "",
   "{0}-of-{1} {2} · {3}/{4} submitted · {5} · {6}": "",
   "{0}/{1} signer keys submitted": "",
@@ -84,6 +85,7 @@ namespace BTCPayServer.Plugins.Translations
   "A payment request with reference ID \"{0}\" already exists for this store.": "",
   "A payment that was made to an approved payout by an external wallet is waiting for your confirmation.": "",
   "A permission to the camera is needed to scan the QR code. Please grant the browser access and then retry.": "",
+  "A plugin install is already queued. Confirming will update the queued install to this version.": "",
   "A Postgres compatible JSON Path (eg. $ ? (@.Customer.Name == \"John\"))": "",
   "A self-hosted, open-source bitcoin payment processor.": "",
   "Abort": "",
@@ -272,6 +274,7 @@ namespace BTCPayServer.Plugins.Translations
   "BTCPay Server currently supports:": "",
   "BTCPay Server could not load the plugin directory.": "",
   "BTCPay Server could not load the plugin panel.": "",
+  "BTCPay Server has queued {0}{1} for this plugin.": "",
   "BTCPay Server Registration will redirect to a custom registration page": "",
   "BTCPay Server sends email requests if email is enabled. You can also share this link or QR code with the selected signers.": "",
   "BTCPay Server Supporters": "",
@@ -292,6 +295,7 @@ namespace BTCPayServer.Plugins.Translations
   "Can use hot wallet": "",
   "Cancel": "",
   "Cancel Invoice": "",
+  "Cancel pending {0}": "",
   "Cannot delete plan. It is currently in use by subscribers.": "",
   "Cannot generate API keys while not using HTTPS or Tor": "",
   "Card reset succeed": "",
@@ -544,6 +548,7 @@ namespace BTCPayServer.Plugins.Translations
   "Delivered at": "",
   "Demonetize": "",
   "Dependencies": "",
+  "Dependencies are not met yet, so BTCPay Server will schedule the install.": "",
   "Dependencies not met.": "",
   "Derivation scheme": "",
   "Description": "",
@@ -568,6 +573,7 @@ namespace BTCPayServer.Plugins.Translations
   "Disable zero amount invoices": "",
   "Disabled": "",
   "Disabled Plugins": "",
+  "Disabled version": "",
   "Disclaimer of Responsibility:": "",
   "Disabling will delete your rate script.": "",
   "Discount": "",
@@ -830,6 +836,7 @@ namespace BTCPayServer.Plugins.Translations
   "I have written down my recovery phrase and stored it in a secure location": "",
   "Id": "",
   "ID": "",
+  "Identifier": "",
   "If a translation isn’t available, it will be searched in the fallback.": "",
   "If authorized, the generated API key will be provided to": "",
   "If BTCPay Server shows you an invalid balance, {0}.<br />If some transactions appear in BTCPay Server, but are missing in another wallet, {1}.": "",
@@ -869,8 +876,10 @@ namespace BTCPayServer.Plugins.Translations
   "Inputs": "",
   "Insert your security device and proceed.": "",
   "Install": "",
+  "Install plugin": "",
   "Installed languages": "",
   "Installed Plugins": "",
+  "Installed version": "",
   "Instructions": "",
   "Interval": "",
   "Invalid": "",
@@ -1270,6 +1279,7 @@ namespace BTCPayServer.Plugins.Translations
   "Payouts Pending": "",
   "Pending": "",
   "Pending Action": "",
+  "Pending action": "",
   "Pending actions": "",
   "Pending Actions": "",
   "Pending Approval": "",
@@ -1323,6 +1333,7 @@ namespace BTCPayServer.Plugins.Translations
   "Please wait for your node to be synched": "",
   "Please, confirm on the device first...": "",
   "Please, enter the passphrase on the device.": "",
+  "Plugin": "",
   "Plugin action cancelled.": "",
   "Plugin Directory": "",
   "Plugin directory": "",
@@ -1980,6 +1991,7 @@ namespace BTCPayServer.Plugins.Translations
   "This page exposes information to use the configured BTCPay Server Configurator to modify this setup.": "",
   "This page is served in non-secure context (HTTPS, localhost or file://)": "",
   "This payment method requires javascript.": "",
+  "This plugin is already installed on the server.": "",
   "This permission is not available for your account.": "",
   "This process disables 2FA until you verify your authenticator app. If you do not complete your authenticator app configuration you may lose access to your account.": "",
   "This plugin cannot be uninstalled as it is depended on by other plugins.": "",
@@ -2192,6 +2204,9 @@ namespace BTCPayServer.Plugins.Translations
   "via HTTPS": "",
   "via TCP or unix domain socket connection": "",
   "via the REST API": "",
+  "Version": "",
+  "Version {0} is available. Manage updates from Installed Plugins.": "",
+  "Version {0} is disabled on this server.": "",
   "View": "",
   "We recommend users exercise caution and conduct their own research or consult the community before installing any plugin.": "",
   "View All": "",
