+++ BTCPayServer/Views/UIServer/PluginDirectory.cshtml
@@ -45,7 +45,7 @@
 {
     <div class="alert alert-info mb-4 d-flex align-items-center justify-content-between">
         <span text-translate="true">You need to restart BTCPay Server in order to update your active plugins.</span>
-        <form method="post" asp-action="Maintenance" class="mt-2">
+        <form method="post" asp-controller="UIServer" asp-action="Maintenance" class="mt-2">
             <button type="submit" name="command" value="soft-restart" class="btn btn-info" text-translate="true">Restart now</button>
         </form>
     </div>
@@ -107,7 +107,7 @@
         </div>
         <div class="offcanvas-body d-flex flex-column overflow-hidden px-4 pt-0 pb-4">
             <div id="selected-plugin-panel" class="d-flex flex-grow-1">
-                <partial name="_SelectedPluginPanel" model="Model.SelectedPluginPanel" />
+                <partial name="~/Views/UIServer/_SelectedPluginPanel.cshtml" model="Model.SelectedPluginPanel" />
             </div>
         </div>
     </div>
